requirements.txt
Pillow
scikit-image
numpy
